add target=_blank in resource verifier link. ♥♥
