+++ README.md
@@ -1,3 +1,5 @@
+[![powered-by-vercel](https://user-images.githubusercontent.com/50735025/117214368-55650a80-ae1a-11eb-8e0c-12d028d25e7e.png)](https://vercel.com/?utm_source=Covidarmy&utm_campaign=oss)
+
 # covid-army - Covid 19 Twitter Resources Aggregator
 
 **covid-army** scrapes tweets from Twitter's search page (using Twitter's advanced search feature) and collates them on one page
