+++ package.json
@@ -4,7 +4,7 @@
   "private": false,
   "license": "MIT",
   "scripts": {
-    "dev": "next dev",
+    "dev": "next dev -p 3007",
     "build": "next build",
     "start": "next start",
     "format": "prettier --write ./src/**/*.{js,jsx,ts,tsx}",
