+++ src/pages/about.tsx
@@ -62,6 +62,7 @@ const About: React.FC<Props> = (props) => {
             <div
               className="flex justify-center w-1/2 p-6 align-middle md:w-1/3 xl:w-1/4"
               key={partner.id}
+              title={partner?.description ?? partner.name}
             >
               <a
                 href={partner.url}
@@ -69,7 +70,7 @@ const About: React.FC<Props> = (props) => {
                 referrerPolicy="no-referrer"
               >
                 <Image
-                  src={`/static/assets/${partner.imageFileName}`}
+                  src={`/static/assets/partners/${partner.imageFileName}`}
                   height={150}
                   width={150}
                 />
